random access works
