--- tests/data/cnr-2000-hc-t.properties
@@ -1,35 +1,0 @@
-#BVGraph properties
-#Fri Nov 19 01:52:04 CET 2010
-compratio=0.131
-bitsforblocks=867390
-residualarcs=492222
-version=0
-zetak=3
-residualexpstats=92858,77484,44429,36518,43311,41369,24344,16645,12746,9854,9189,8229,6673,4725,2522,1028,1862,2866,2252,514
-avgref=47.473
-residualavggap=5254.925
-avgbitsforoutdegrees=4.254
-windowsize=16
-bitsforintervals=552694
-copiedarcs=2282054
-avgbitsforblocks=2.664
-bitsperlink=2.157
-bitsforresiduals=3432508
-bitsforreferences=698450
-avgdist=1.145
-successoravgloggap=2.0091557973334147
-avgbitsforreferences=2.145
-successoravggap=1021.639
-maxrefcount=2147483647
-successorexpstats=2397257,180185,104571,97134,101425,86303,56050,37444,28910,26883,23821,15177,10972,7260,4508,1418,2727,4115,2959,690
-nodes=325557
-compressionflags=
-intervalisedarcs=441876
-bitspernode=21.305
-arcs=3216152
-bitsforoutdegrees=1385035
-avgbitsforintervals=1.698
-minintervallength=4
-graphclass=it.unimi.dsi.webgraph.BVGraph
-avgbitsforresiduals=10.543
-residualavgloggap=4.4034610567162
