--- tests/data/cnr-2000-hc.properties
@@ -1,35 +1,0 @@
-#BVGraph properties
-#Fri Nov 19 01:51:55 CET 2010
-compratio=0.149
-bitsforblocks=1335385
-residualarcs=472175
-version=0
-zetak=3
-residualexpstats=85112,72580,35560,28157,36401,25945,19435,15322,14741,15002,15978,15943,14695,10071,5090,1445,2527,3673,2736,550
-avgref=300.979
-residualavggap=7096.588
-avgbitsforoutdegrees=5.1
-windowsize=16
-bitsforintervals=466355
-copiedarcs=2487905
-avgbitsforblocks=4.102
-bitsperlink=2.448
-bitsforresiduals=3729558
-bitsforreferences=682582
-avgdist=1.336
-successoravgloggap=2.6133784090933467
-avgbitsforreferences=2.097
-successoravggap=1803.755
-maxrefcount=2147483647
-successorexpstats=2232567,192413,101723,78903,86825,66745,59116,49458,51851,60901,55054,50966,48420,22077,12427,3153,4827,7478,4679,774
-nodes=325557
-compressionflags=
-intervalisedarcs=256072
-bitspernode=24.187
-arcs=3216152
-bitsforoutdegrees=1660205
-avgbitsforintervals=1.432
-minintervallength=4
-graphclass=it.unimi.dsi.webgraph.BVGraph
-avgbitsforresiduals=11.456
-residualavgloggap=5.186772989517994
