+++ tests/data/cnr-2000.properties
@@ -1,0 +1,35 @@
+#BVGraph properties
+#Wed Nov 17 23:14:11 CET 2010
+compratio=0.176
+bitsforblocks=1353080
+residualarcs=577350
+version=0
+zetak=3
+residualexpstats=108927,84647,44166,36932,46551,32420,25074,19244,19463,20741,21880,20784,18799,13237,6404,1705,2875,4232,3124,611
+avgref=1.311
+residualavggap=6531.255
+avgbitsforoutdegrees=5.1
+windowsize=7
+bitsforintervals=829187
+copiedarcs=2195145
+avgbitsforblocks=4.156
+bitsperlink=2.897
+bitsforresiduals=4694729
+bitsforreferences=781540
+avgdist=1.64
+successoravgloggap=2.6133784090933467
+avgbitsforreferences=2.401
+successoravggap=1803.755
+maxrefcount=3
+successorexpstats=2232567,192413,101723,78903,86825,66745,59116,49458,51851,60901,55054,50966,48420,22077,12427,3153,4827,7478,4679,774
+nodes=325557
+compressionflags=
+intervalisedarcs=443657
+bitspernode=28.624
+arcs=3216152
+bitsforoutdegrees=1660205
+avgbitsforintervals=2.547
+minintervallength=4
+graphclass=it.unimi.dsi.webgraph.BVGraph
+avgbitsforresiduals=14.421
+residualavgloggap=5.246773838359595
