+++ tests/data/cnr-2000_edges.txt
@@ -1,0 +1,55 @@
+1 4 8 219 220 
+0 7 8 219 220 
+3 4 8 219 220 
+2 8 9 219 220 
+0 2 8 219 220 
+6 8 45 219 220 
+5 7 8 219 220 
+1 6 8 219 220 
+0 1 2 3 4 5 6 7 9 10 11 12 13 14 54 64 146 156 
+3 8 10 219 220 
+8 9 11 219 220 
+8 10 13 219 220 
+8 13 14 219 220 
+8 11 12 219 220 
+8 12 99 219 220 
+16 17 18 19 20 21 22 23 24 25 26 27 28 29 64 76 146 156 
+15 17 26 219 220 
+15 16 18 219 220 
+15 17 19 219 220 
+15 18 20 219 220 
+15 19 21 219 220 
+15 20 22 219 220 
+15 21 23 219 220 
+15 22 24 219 220 
+15 23 27 219 220 
+15 26 28 219 220 
+15 16 25 219 220 
+15 24 71 219 220 
+15 25 29 219 220 
+15 28 61 219 220 
+31 39 208 219 220 
+30 32 39 219 220 
+31 33 39 219 220 
+32 34 39 219 220 
+33 35 39 219 220 
+34 36 39 219 220 
+35 37 39 219 220 
+36 38 39 219 220 
+37 39 40 219 220 
+30 31 32 33 34 35 36 37 38 40 41 42 43 44 54 146 156 206 
+38 39 41 219 220 
+39 40 42 219 220 
+39 41 43 219 220 
+39 42 44 219 220 
+39 43 59 219 220 
+5 48 54 219 220 
+47 54 55 219 220 
+46 48 54 219 220 
+45 47 54 219 220 
+50 54 56 219 220 
+49 51 54 219 220 
+50 52 54 219 220 
+51 53 54 219 220 
+52 54 55 219 220 
+8 39 45 46 47 48 49 50 51 52 53 55 56 57 58 59 146 156
